compress binary
